requirements.txt
openai~=1.13.3
tenacity~=8.2.3
termcolor~=2.4.0
requests~=2.31.0
python-dotenv~=1.0.1
Pyrogram~=2.0.106